package: add :upgrade action
Since there is no generic way of checking whether there is a new version
of a given package, the upgrade action just installs the package
unconditionally, assuming that the package installation command will
install a newer version if it's available (that's the case for e.g.
apt-get).

Since it's only possible to check whether there was any version change
_after_ the action, show_differences needs to be called explicitly when
the package version changes for us to get some feedback in the logs.